requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.1.0
python-dateutil==2.8.2